Fix a bug with @content_for
Since `render` will render whatever it's given and puke the output into
the current context, we need to give @content_for a function which will
render the appropriate template so it can be run when the time's right.

The reason why this worked so far is that the `footer` view was empty,
so nothing got rendered, and the `nav` view, coincidentally, got
rendered at just the right place, so everything seemed normal.