Add CSRF protection
Using the lapis.csrf module, now added a hidden CSRF token to the form
in the off chance this might get exploited. Also wrapped
submit-application call in `pcall` in case it throws a Lua error, so we
are able to catch it and send a JSON response telling the client
something's wrong, instead of defaulting to an HTML page.